cap9-cli: Restructure deployment.
