+++ kernel-ewasm/cap9-cli/src/fetch/utils.rs
@@ -7,8 +7,8 @@ use web3::Transport;
 use rustc_hex::FromHex;
 use rustc_hex::ToHex;
 // use ethabi::Token::Uint;
-use crate::conn;
-use crate::conn::EthConn;
+use crate::connection;
+use crate::connection::EthConn;
 use crate::project::LocalProject;
 use cap9_std::proc_table::cap::*;
 use pwasm_abi;
@@ -20,7 +20,7 @@ use cap9_core::*;
 use cap9_core::Error;
 use cap9_core::Read;
 use crate::constants;
-use crate::deploy::{from_common_u256, to_common_u256, to_common_h256,
+use crate::utils::{from_common_u256, to_common_u256, to_common_h256,
     from_common_address, to_common_address
 };
 use std::collections::{HashMap, HashSet};
